move upgrade to setup
